+++ scripts/index.js
@@ -3,22 +3,27 @@
     some general and other stuff that has no place yet
     also includes unused stuff that i may consider in the future */
 
-$('#watermark').text(`${watermark} | `);
+$('#watermark').text(`${watermark}`);
 $('#version').text(version);
 
 // ui stuff
 if (flag.tool.dd) $('#dd').prop('checked', true);
-if (flag.tool.hb.staircase) $('#hb-stair').prop('checked', true);
-if (flag.tool.vb.note) $('#vb-note').prop('checked', true);
+if (flag.tool.hbStaircase) $('#hb-stair').prop('checked', true);
+if (flag.tool.vbSpecific === 'time') $('#vb-specific-time').prop('checked', true);
+if (flag.tool.vbSpecific === 'diff') $('#vb-specific-diff').prop('checked', true);
+if (flag.tool.vbNote) $('#vb-note').prop('checked', true);
 if (flag.tool.shrAngle) $('#shrangle').prop('checked', true);
 if (flag.tool.speedPause) $('#speedpause').prop('checked', true);
+if (flag.tool.slowSlider) $('#slowslider').prop('checked', true);
 
 $('#ebpm').val(tool.ebpm.th);
 $('#ebpms').val(tool.ebpm.thSwing);
+$('#slowslider-min').val(tool.minSliderSpeed * 1000);
+$('#slowslider-min-prec').val(0);
 $('#shrangle-max').val(tool.maxShrAngle * 1000);
-$('#shrangle-max-beat').val(0);
+$('#shrangle-max-prec').val(0);
 $('#speedpause-max').val(tool.maxSpeedPause * 1000);
-$('#speedpause-max-beat').val(0);
+$('#speedpause-max-prec').val(0);
 $('#vb-min').val(tool.vb.min * 1000);
 $('#vb-min-beat').val(0);
 $('#vb-max').val(tool.vb.max * 1000);
@@ -38,6 +43,20 @@ if ($.urlParam('url') !== null) {
     downloadFromURL($.urlParam('url'));
 } else if ($.urlParam('id') !== null) {
     downloadFromID($.urlParam('id'));
+}
+// used for debugging, skipping audio process
+if ($.urlParam('noaudio') !== null) {
+    flag.noAudio = $.urlParam('noaudio') === 'true' ? true : false;
+}
+if ($.urlParam('nostats') !== null) {
+    flag.noStats = $.urlParam('nostats') === 'true' ? true : false;
+    $('#stats').css({ display: 'none' });
+}
+if ($.urlParam('tools') !== null) {
+    flag.noStats = $.urlParam('tools') === 'true' ? true : false;
+    setTimeout(function () {
+        $('.accordion').trigger('click');
+    }, 100);
 }
 
 function toMMSS(num) {
@@ -92,11 +111,9 @@ function offsetBegone(beat, bpm, offset) {
 
 // just to make rounding with decimal easier
 function round(num, d = 0) {
-    if (!d > 0) {
-        return Math.round(num);
-    }
+    const place = Math.pow(10, d);
     // return parseFloat(num.toFixed(d));
-    return Math.round(num * Math.pow(10, d)) / Math.pow(10, d);
+    return Math.round(num * place) / place;
 }
 
 // thanks Top_Cat#1961