+++ scripts/tools/note.js
@@ -5,30 +5,26 @@
 
 // smort
 function getNoteCount(notes) {
-    let nr = 0;
-    let nb = 0;
-    let b = 0;
-    let cn = 0;
-    let cb = 0;
+    const noteCount = { red: 0, blue: 0, bomb: 0, chromaN: 0, chromaB: 0 };
     for (let i = notes.length - 1; i >= 0; i--) {
         if (notes[i]._type === 0) {
-            nr++;
+            noteCount.red++;
             if (notes[i]._type._customData?._color) {
-                cn++;
+                noteCount.chromaN++;
             }
         } else if (notes[i]._type === 1) {
-            nb++;
+            noteCount.blue++;
             if (notes[i]._type._customData?._color) {
-                cn++;
+                noteCount.chromaN++;
             }
         } else if (notes[i]._type === 3) {
-            b++;
+            noteCount.bomb++;
             if (notes[i]._type._customData?._color) {
-                cb++;
+                noteCount.chromaB++;
             }
         }
     }
-    return { red: nr, blue: nb, bomb: b, chromaN: cn, chromaB: cb };
+    return noteCount;
 }
 function countNoteLayer(notes, l) {
     let count = 0;
@@ -166,76 +162,62 @@ function getEffectiveBPMSwingTime(diff, mapSettings) {
     return arr;
 }
 function getMinSliderSpeed(notes) {
-    let speed = 0;
+    let speedR = 0;
+    let speedB = 0;
     let lastRed;
     let lastBlue;
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (note._type === 0) {
             if (lastRed) {
-                if (swingNext(note, lastRed)) {
-                    lastRed = note;
-                } else if (swingWindow(note, lastRed)) {
-                    speed = Math.max(speed, toRealTime(note._time - lastRed._time) / 2);
-                } else {
-                    speed = Math.max(speed, toRealTime(note._time - lastRed._time));
+                if (!swingNext(note, lastRed)) {
+                    speedR = Math.max(speedR, toRealTime(note._time - lastRed._time) / (swingWindow(note, lastRed) ? 2 : 1));
                 }
+                lastRed = note;
             } else {
                 lastRed = note;
             }
         } else if (note._type === 1) {
             if (lastBlue) {
-                if (swingNext(note, lastBlue)) {
-                    lastBlue = note;
-                } else if (swingWindow(note, lastBlue)) {
-                    speed = Math.max(speed, toRealTime(note._time - lastBlue._time) / 2);
-                } else {
-                    speed = Math.max(speed, toRealTime(note._time - lastBlue._time));
+                if (!swingNext(note, lastBlue)) {
+                    speedB = Math.max(speedB, toRealTime(note._time - lastBlue._time) / (swingWindow(note, lastBlue) ? 2 : 1));
                 }
+                lastBlue = note;
             } else {
                 lastBlue = note;
             }
         }
     }
-    return speed;
+    return Math.max(speedR, speedB);
 }
 function getMaxSliderSpeed(notes) {
-    let speed = Number.MAX_SAFE_INTEGER;
+    let speedR = Number.MAX_SAFE_INTEGER;
+    let speedB = Number.MAX_SAFE_INTEGER;
     let lastRed;
     let lastBlue;
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (note._type === 0) {
             if (lastRed) {
-                if (swingNext(note, lastRed)) {
-                    lastRed = note;
-                } else if (toRealTime(note._time - lastRed._time) > 0.001) {
-                    if (swingWindow(note, lastRed)) {
-                        speed = Math.min(speed, toRealTime(note._time - lastRed._time) / 2);
-                    } else {
-                        speed = Math.min(speed, toRealTime(note._time - lastRed._time));
-                    }
+                if (!swingNext(note, lastRed) && toRealTime(note._time - lastRed._time) > 0.001) {
+                    speedR = Math.min(speedR, toRealTime(note._time - lastRed._time) / (swingWindow(note, lastRed) ? 2 : 1));
                 }
+                lastRed = note;
             } else {
                 lastRed = note;
             }
         } else if (note._type === 1) {
             if (lastBlue) {
-                if (swingNext(note, lastBlue)) {
-                    lastBlue = note;
-                } else if (toRealTime(note._time - lastBlue._time) > 0.001) {
-                    if (swingWindow(note, lastBlue)) {
-                        speed = Math.min(speed, toRealTime(note._time - lastBlue._time) / 2);
-                    } else {
-                        speed = Math.min(speed, toRealTime(note._time - lastBlue._time));
-                    }
+                if (!swingNext(note, lastBlue) && toRealTime(note._time - lastBlue._time) > 0.001) {
+                    speedB = Math.min(speedB, toRealTime(note._time - lastBlue._time) / (swingWindow(note, lastBlue) ? 2 : 1));
                 }
+                lastBlue = note;
             } else {
                 lastBlue = note;
             }
         }
     }
-    return speed;
+    return Math.min(speedR, speedB);
 }
 
 function detectDoubleDirectional(diff, mapSettings) {
@@ -350,29 +332,55 @@ function checkDD(n1cd, n2cd) {
 
 function detectVisionBlock(diff, mapSettings) {
     const { _notes: notes } = diff;
-    const { bpm, bpmc, offset } = mapSettings;
+    const { diffName, hjd, bpm, bpmc, offset } = mapSettings;
     const arr = [];
     let lastMidL;
     let lastMidR;
     for (let i = 0, len = notes.length; i < len; i++) {
         const note = notes[i];
         if (lastMidL) {
-            if (isAboveTH(note._time - lastMidL._time, tool.vb.min) && isBelowTH(note._time - lastMidL._time, tool.vb.max)) {
+            if (
+                isAboveTH(
+                    note._time - lastMidL._time,
+                    flag.tool.vbSpecific === 'time' ? tool.vb.min : tool.vb.diff[diffName].min
+                ) &&
+                isBelowTH(
+                    note._time - lastMidL._time,
+                    flag.tool.vbSpecific === 'time' ? tool.vb.max : Math.min(toRealTime(hjd), tool.vb.diff[diffName].max)
+                )
+            ) {
                 if (note._lineIndex < 2) {
                     arr.push(adjustTime(note._time, bpm, offset, bpmc));
                 }
             }
             // yeet the last note if nothing else found so we dont have to perform check every note
-            else if (toRealTime(note._time - lastMidL._time) >= tool.vb.max) {
+            else if (
+                toRealTime(note._time - lastMidL._time) >= flag.tool.vbSpecific === 'time'
+                    ? tool.vb.max
+                    : Math.min(toRealTime(hjd), tool.vb.diff[diffName].max)
+            ) {
                 lastMidL = null;
             }
         }
         if (lastMidR) {
-            if (isAboveTH(note._time - lastMidR._time, tool.vb.min) && isBelowTH(note._time - lastMidR._time, tool.vb.max)) {
+            if (
+                isAboveTH(
+                    note._time - lastMidR._time,
+                    flag.tool.vbSpecific === 'time' ? tool.vb.min : tool.vb.diff[diffName].min
+                ) &&
+                isBelowTH(
+                    note._time - lastMidR._time,
+                    flag.tool.vbSpecific === 'time' ? tool.vb.max : Math.min(toRealTime(hjd), tool.vb.diff[diffName].max)
+                )
+            ) {
                 if (note._lineIndex > 1) {
                     arr.push(adjustTime(note._time, bpm, offset, bpmc));
                 }
-            } else if (toRealTime(note._time - lastMidR._time) >= tool.vb.max) {
+            } else if (
+                toRealTime(note._time - lastMidR._time) >= flag.tool.vbSpecific === 'time'
+                    ? tool.vb.max
+                    : Math.min(toRealTime(hjd), tool.vb.diff[diffName].max)
+            ) {
                 lastMidR = null;
             }
         }
@@ -819,4 +827,54 @@ function detectImproperWindowSnap(diff, mapSettings) {
 
 function isSlantedWindow(n1, n2) {
     return swingWindow(n1, n2) && !swingDiagonal(n1, n2) && !swingHorizontal(n1, n2) && !swingVertical(n1, n2);
+}
+
+function detectSlowSlider(diff, mapSettings) {
+    const { _notes: notes } = diff;
+    const { bpm, bpmc, offset } = mapSettings;
+    const arr = [];
+    let speedR = 0;
+    let speedB = 0;
+    let lastRedTime;
+    let lastBlueTime;
+    let lastRed;
+    let lastBlue;
+    for (let i = 0, len = notes.length; i < len; i++) {
+        const note = notes[i];
+        if (note._type === 0) {
+            if (lastRed) {
+                if (swingNext(note, lastRed)) {
+                    speedR = 0;
+                    lastRedTime = note._time;
+                } else {
+                    speedR = Math.max(speedR, toRealTime(note._time - lastRed._time) / (swingWindow(note, lastRed) ? 2 : 1));
+                }
+                lastRed = note;
+                if (speedR > tool.minSliderSpeed) {
+                    arr.push(adjustTime(lastRedTime, bpm, offset, bpmc));
+                }
+            } else {
+                lastRed = note;
+            }
+        }
+        if (note._type === 1) {
+            if (lastBlue) {
+                if (swingNext(note, lastBlue)) {
+                    speedB = 0;
+                    lastBlueTime = note._time;
+                } else {
+                    speedB = Math.max(speedB, toRealTime(note._time - lastBlue._time) / (swingWindow(note, lastBlue) ? 2 : 1));
+                }
+                lastBlue = note;
+                if (speedB > tool.minSliderSpeed) {
+                    arr.push(adjustTime(lastBlueTime, bpm, offset, bpmc));
+                }
+            } else {
+                lastBlue = note;
+            }
+        }
+    }
+    return arr.filter(function (x, i, ary) {
+        return !i || x !== ary[i - 1];
+    });
 }